+++ index.js
@@ -31,32 +31,19 @@ const boomFunctions = [
   'serverUnavailable',
   'gatewayTimeout',
   'illegal',
-  'teapot'
+  'teapot',
+  'boomify'
 ]
 
-module.exports.register = (server, options, next) => {
+const register = async (server, options) => {
   boomFunctions.forEach(boomFunction => {
-    server.decorate('reply', boomFunction, function () {
-      this.response(Boom[boomFunction].apply(Boom, arguments))
+    server.decorate('toolkit', boomFunction, function () {
+      throw Boom[boomFunction].apply(Boom, arguments)
     })
   })
-
-  server.decorate('reply', 'boom', function () {
-    var args = Array.prototype.slice.call(arguments)
-    let boom
-
-    if (args.length > 1 && args[1] instanceof Error) {
-      boom = Boom.wrap(args[1], args[0], args.slice(2, args.length - 2))
-    } else {
-      boom = Boom.create.apply(null, args)
-    }
-
-    this.response(boom)
-  })
-
-  next()
 }
 
-module.exports.register.attributes = {
+module.exports.plugin = {
+  register,
   pkg: require('./package.json')
 }